backend/requirements.txt
Django
django-cors-headers
djangorestframework
pandas
numpy
scikit-learn
joblib
ee
geopy
windows-curses
earthengine-api
celery
redis
